Close #39
